fix: change wrap/unwrap parameter from uint32 to uint128
The encrypted balance type is euint128 (128-bit), but wrap() and unwrap()
accepted uint32 parameters, capping single-operation amounts at 2^32-1
(~4.3B). For tokens with 18 decimals this means a maximum of ~4.29 tokens
per wrap/unwrap call. Changing to uint128 aligns the plaintext parameter
with the underlying encrypted type capacity.

+++ contracts/experimental/token/FHERC20/FHERC20.sol
@@ -69,7 +69,7 @@ contract FHERC20 is IFHERC20, ERC20, Permissioned {
         return _transferImpl(from, to, spent);
     }
 
-    function wrap(uint32 amount) public {
+    function wrap(uint128 amount) public {
         if (balanceOf(msg.sender) < amount) {
             revert ErrorInsufficientFunds();
         }
@@ -80,7 +80,7 @@ contract FHERC20 is IFHERC20, ERC20, Permissioned {
         totalEncryptedSupply = totalEncryptedSupply + eAmount;
     }
 
-    function unwrap(uint32 amount) public {
+    function unwrap(uint128 amount) public {
         euint128 encAmount = FHE.asEuint128(amount);
 
         euint128 amountToUnwrap = FHE.select(_encBalances[msg.sender].gte(encAmount), encAmount, FHE.asEuint128(0));
